tambahin handler jika id tidak ada di database

+++ service/barang.go
@@ -1,6 +1,7 @@
 package service
 
 import (
+	"database/sql"
 	"fmt"
 	_ "github.com/go-sql-driver/mysql"
 	"log"
@@ -64,6 +65,8 @@ func Edit_barang(){
 	fmt.Println("Masukkan id barang")
 	fmt.Scan(&id_barang)
 
+	if FindBarang(id_barang) {
+
 	fmt.Println("Masukkan nama barang")
 	fmt.Scan(&nama_barang)
 
@@ -76,12 +79,23 @@ func Edit_barang(){
 	fmt.Println("Masukkan id_satuan")
 	fmt.Scan(&id_satuan)
 
-	update,err := db.Prepare("UPDATE barang SET nama_barang=?, stok=?, id_kategori =?, id_satuan =? WHERE id_barang = ?")
-	if err != nil {
-		log.Fatalf("Terjadi error terkait edit supplier karena: ", err)
+
+		update,err := db.Prepare("UPDATE barang SET nama_barang=?, stok=?, id_kategori =?, id_satuan =? WHERE id_barang = ?")
+		if err != nil {
+			log.Fatalf("Terjadi error terkait edit supplier karena: ", err)
+		}
+		if err == sql.ErrNoRows{
+			log.Fatalln("Row tidak ditemukan di fungsi Edit barang karena")
+		}
+	
+		update.Exec(nama_barang,stok,id_kategori, id_satuan, id_barang)
+	} else {
+		fmt.Println("Id tidak ada")
+	
+
+
 	}
 
-	update.Exec(nama_barang,stok,id_kategori, id_satuan, id_barang)
 }
 
 func ShowAllBarang(){
@@ -92,6 +106,7 @@ func ShowAllBarang(){
 	if err != nil {
 		log.Fatalf("Terjadi error terkait Show All Supplier karena: ", err)
 	}
+
 
 	bar:= Barang{}
 	barList:= []Barang{}
@@ -133,33 +148,41 @@ func ShowPerBarang(){
 	fmt.Println("Masukkan id barang")
 	fmt.Scan(&id_barang)
 
-	selDB,err := db.Query("SELECT * FROM barang WHERE id_barang = ?",id_barang)
-	if err != nil {
-		log.Fatalf("Terjadi error terkait query barang by id karena error: ", err)
-	}
-
-	bar:= Barang{}
-
-	for selDB.Next(){
-		var id_barang1 int
-		var nama_barang string
-		var stok int
-		var id_kategori int
-		var id_satuan int
-
-		err = selDB.Scan(&id_barang1, &nama_barang, &stok, &id_kategori, &id_satuan)
+	if FindBarang(id_barang) {
+		selDB,err := db.Query("SELECT * FROM barang WHERE id_barang = ?",id_barang)
 		if err != nil {
-			log.Fatalf("Terjadi error dikarenakan scan barang by id", err)
+			log.Fatalf("Terjadi error terkait query barang by id karena error: ", err)
 		}
-
-		bar.Id_barang = id_barang1
-		bar.Nama_barang = nama_barang
-		bar.Stok = stok
-		bar.Id_kategori = id_kategori
-		bar.Id_satuan = id_satuan
+	
+			bar:= Barang{}
+	
+			for selDB.Next(){
+				var id_barang1 int
+				var nama_barang string
+				var stok int
+				var id_kategori int
+				var id_satuan int
+		
+				err = selDB.Scan(&id_barang1, &nama_barang, &stok, &id_kategori, &id_satuan)
+				if err != nil {
+					log.Fatalf("Terjadi error dikarenakan scan barang by id", err)
+				}
+		
+				bar.Id_barang = id_barang1
+				bar.Nama_barang = nama_barang
+				bar.Stok = stok
+				bar.Id_kategori = id_kategori
+				bar.Id_satuan = id_satuan
+			}
+		
+			fmt.Println(bar)
+	} else {
+		fmt.Println("Id tidak ada")
 	}
 
-	fmt.Println(bar)
+
+	
+
 }
 
 func Delete_barang(){
@@ -171,13 +194,45 @@ func Delete_barang(){
 
 	fmt.Scan(&id_barang)
 
-	
-	result,err := db.Exec("DELETE FROM barang WHERE id_barang=?", id_barang)
+	is_exist:= FindBarang(id_barang)
 
-	if err!= nil {
-		log.Fatalln("Terjadi error terkait delForm", err)
+	if is_exist == true {
+		result,err := db.Exec("DELETE FROM barang WHERE id_barang=?", id_barang)
+
+		if err!= nil {
+			log.Fatalln("Terjadi error terkait delForm", err)
+		}
+		if err == sql.ErrNoRows{
+			log.Fatalln("Row tidak ditemukan di fungsi ShowPerBarang karena")
+		}
+		fmt.Println(result)
+	} else {
+		fmt.Println("id tidak ada di database")
 	}
-	fmt.Println(result)
+	
 
 
+
+}
+
+func FindBarang(id_barang int) bool{
+	db:=dbConn()
+	defer db.Close()
+
+	row := db.QueryRow("SELECT * FROM barang WHERE id_barang = ?", id_barang)
+
+	bar := Barang{}
+	err:= row.Scan(&bar.Id_barang, &bar.Nama_barang, &bar.Stok, &bar.Id_kategori, &bar.Id_satuan)
+
+	switch {
+	case err == sql.ErrNoRows:
+		// log.Fatalf("Barang dari id yang dimasukkan tidak ada di database")
+		return false
+	case err != nil:
+		log.Fatalf("Error saat mencari barang karena : ", err)
+		return false
+	}
+
+	fmt.Println(bar)
+	return true
 }
